cập nhật và thêm phân quyền người dùng

+++ app/Http/Controllers/Admin/AdminController.php
@@ -9,6 +9,7 @@ use App\Models\Admin;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Session;
+use Spatie\Permission\Models\Role;
 
 class AdminController extends Controller
 {
@@ -17,10 +18,9 @@ class AdminController extends Controller
      */
     public function index()
     {
-        //
-
+        $this->authorize('xem-nhan-vien');
         $title = 'Danh sách nhân viên';
-        $managers = Admin::orderByDesc('id')->paginate(15);
+        $managers = Admin::with('role')->orderByDesc('id')->paginate(15);
         return view('admin.manager.list', compact('title', 'managers'));
     }
 
@@ -29,8 +29,10 @@ class AdminController extends Controller
      */
     public function create()
     {
+        $this->authorize('them-nhan-vien');
         $title = 'Thêm mới nhân viên';
-        return view('admin.manager.create', compact('title'));
+        $roles = Role::all();
+        return view('admin.manager.create', compact('title', 'roles'));
     }
 
     /**
@@ -38,12 +40,14 @@ class AdminController extends Controller
      */
     public function store(AdminRequest $request)
     {
+        $this->authorize('them-nhan-vien');
         try {
             $rand = rand(100000, 999999);
             $admin = Admin::create([
                 'name' => $request->name,
                 'email' => $request->email,
-                'password' => Hash::make($rand)
+                'password' => Hash::make($rand),
+                'role_id' => $request->role_id,
             ]);
             Session::flash('success', 'Tạo nhân viên thành công');
             MailAccountJob::dispatch($admin, $rand)->delay(now()->addSecond(5));
@@ -66,10 +70,11 @@ class AdminController extends Controller
      */
     public function edit(string $id)
     {
-        
+        $this->authorize('sua-nhan-vien');
         $manager = Admin::findOrFail($id);
         $title = 'Chỉnh sửa người quản lý';
-        return view('admin.manager.edit', compact('title', 'manager'));
+        $roles = Role::all();
+        return view('admin.manager.edit', compact('title', 'manager', 'roles'));
     }
 
     /**
@@ -77,25 +82,34 @@ class AdminController extends Controller
      */
     public function update(AdminRequest $request, string $id)
     {
+        $this->authorize('sua-nhan-vien');
         $manager = Admin::findOrFail($id);
+
         try {
             $manager->fill([
                 'name' => $request->name,
                 'email' => $request->email,
                 'phone' => $request->phone,
                 'address' => $request->address,
-                'gender' => $request->gender
+                'gender' => $request->gender,
+                'role_id' => $request->role_id,
             ]);
+
             if ($request->password) {
                 $manager->password = Hash::make($request->password);
             }
+
             $manager->save();
+
+            $role = Role::findById($request->role_id);
+            $manager->syncRoles([$role->name]); 
+
             Session::flash('success', 'Cập nhật nhân viên thành công');
             return redirect()->route('manager.index');
         } catch (\Exception $e) {
             Session::flash('error', 'Có lỗi khi chỉnh sửa: ' . $e->getMessage());
+            return redirect()->back()->withInput();
         }
-        return redirect()->back();
     }
 
     /**
@@ -103,9 +117,16 @@ class AdminController extends Controller
      */
     public function destroy(string $id)
     {
+        $this->authorize('xoa-nhan-vien');
         $manager = Admin::find($id);
         if (!$manager) {
-            abort('404');
+            abort(404);
+        }
+        if ($manager->hasRole('superadmin')) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Bạn không có quyền xóa tài khoản Superadmin.'
+            ]);
         }
         try {
             $manager->delete();
